requirements.txt
dnspython==2.4.1
gssapi==1.8.2
ldap3==2.9.1
paramiko==3.2.0
